sadik-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
pyserial>=3.5
websockets>=12.0
httpx>=0.26.0
openai>=1.12.0
edge-tts>=6.1.9
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
